Add explicit type conversion

+++ lib/stages/FEngine.cpp
@@ -1028,7 +1028,7 @@ void FEngine::main_thread() {
                 for (int n = 0; n < num_dish_locations_ns * num_dish_locations_ew
                                         * num_polarizations * num_local_channels * U;
                      ++n)
-                    ((std::complex<float16_t>*)W1_frame)[n] = frb1_input_scale;
+                    ((std::complex<float16_t>*)W1_frame)[n] = (float16_t)frb1_input_scale;
             }
             DEBUG("[{:d}] Done filling W1 buffer for U={:d}.", W1_frame_index, U);
